Tiny first step at restructuring addons in web
