+++ www/src/Data/Addons.ts
@@ -1,0 +1,21 @@
+export const BUTTON_MASKS = [
+	{ label: 'None', value: 0 },
+	{ label: 'B1', value: 1 << 0 },
+	{ label: 'B2', value: 1 << 1 },
+	{ label: 'B3', value: 1 << 2 },
+	{ label: 'B4', value: 1 << 3 },
+	{ label: 'L1', value: 1 << 4 },
+	{ label: 'R1', value: 1 << 5 },
+	{ label: 'L2', value: 1 << 6 },
+	{ label: 'R2', value: 1 << 7 },
+	{ label: 'S1', value: 1 << 8 },
+	{ label: 'S2', value: 1 << 9 },
+	{ label: 'L3', value: 1 << 10 },
+	{ label: 'R3', value: 1 << 11 },
+	{ label: 'A1', value: 1 << 12 },
+	{ label: 'A2', value: 1 << 13 },
+	{ label: 'Up', value: 1 << 16 },
+	{ label: 'Down', value: 1 << 17 },
+	{ label: 'Left', value: 1 << 18 },
+	{ label: 'Right', value: 1 << 19 },
+];
